Split release helper modules
Pin the `cross` installer to the reviewed commit revision and split the
release helper into focused CLI, version, and packaging modules so each
file stays under the project size limit.

Parse the Cargo manifest in the binstall metadata test instead of matching
raw TOML substrings, preserving the existing release behaviour while making
metadata assertions structural.
