+++ .github/workflows/release.yml
@@ -58,7 +58,7 @@ jobs:
       - name: Install cross
         env:
           RUSTFLAGS: ""
-        run: cargo install cross --git https://github.com/cross-rs/cross --tag v0.2.5
+        run: cargo install cross --git https://github.com/cross-rs/cross --rev 88f49ff79e777bef6d3564531636ee4d3cc2f8d2
       - name: Cache cargo registry
         uses: actions/cache@0057852bfaa89a56745cba8c7296529d2fc39830
         with:
