+++ docs/developers-guide.md
@@ -18,17 +18,18 @@ rerun replaces existing files instead of failing on duplicates.
 
 ## Release helper script
 
-Release version checks, artefact copying, archive generation, and checksum
-creation live in `scripts/release_support.py`. Keep this logic in the script
-rather than embedding it directly in workflow shell blocks so it remains
-testable.
+Release version checks live in `scripts/release_version.py`. Artefact copying,
+archive generation, and checksum creation live in `scripts/release_packaging.py`.
+The `scripts/release_support.py` script keeps the command-line wiring thin.
+Keep this logic in scripts rather than embedding it directly in workflow shell
+blocks so it remains testable.
 
 The script follows the repository scripting standards:
 
 - `uv` executes the script with its declared Python dependencies.
 - Cyclopts maps `INPUT_*` workflow environment variables to typed parameters.
-- Pure helper functions handle version resolution, tag validation, archive
-  creation, and checksum manifest creation.
+- Pure helper functions in focused modules handle version resolution, tag
+  validation, archive creation, and checksum manifest creation.
 
 Run the script tests with:
